refactor(bots): extract SSE chat streaming out of GaiaClient
GaiaClient was a 1027-line god-object; ~340 of those lines were the SSE
streaming machinery (request, incremental parsing, inactivity/keepalive,
session-token capture, transient-error retry). Move it into a focused
api/chat-stream.ts as streamChat(), driven by a small ChatStreamClient deps
interface so it stays decoupled from the client's privates. GaiaClient.chatStream
now delegates; its public signature is unchanged (no call-site churn). index.ts
drops to 724 lines. Verified: all 4 bots type-check and a live Telegram chat
still streams correctly.

Entire-Checkpoint: f38175424fb2
